Remove backend graph-type leakage from IR/compiler types
